requirements.txt
Mako==1.0.3
MarkupSafe==0.23
PasteDeploy==1.5.2
pyramid==1.5.7
pyramid-mako==1.0.2
pyramid-tm==0.12.1
repoze.lru==0.6
translationstring==1.3
venusian==1.0
waitress==0.8.10
WebOb==1.5.1
wsgiref==0.1.2
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.5
-e .